Inline incident reconciliation into validator activity status
